requirements.txt
pyarrow
pandas
apache-beam[gpc]==2.55.0
psycopg2-binary==2.9.9
google-cloud-storage
setuptools
apache_beam[dataframe]
fsspec
gcsfs
google-cloud-secret-manager
pandas-gbq 
google-cloud-bigquery
